compiles, gotta figure out the buckets
works just fine with our player structs, when we figure out how the buckets will be organized, i gotta change some parameters that are acting as placeholders right now. I could add some new functions if we think of any. also the main in this is just me testin it out, not the actual main